Frogger: Move Frog and Enemy classes into separate files.
